Create Re-generate-certs.sh
This script continues to take the hostname and number of days for the certificate to be valid as inputs. It uses certutil from NSS library, which is more secure than openssl to generate the certificate. It then encrypts the private key with a passphrase supplied by the user and creates a PKCS#12 file that contains both the certificate and the encrypted private key. The PKCS#12 file is also protected with a passphrase supplied by the user. After that, the unencrypted private key is deleted and the ownership and permissions of the generated files are adjusted to provide an additional layer of security.

It is important to securely store the PKCS#12 file and passphrases, and to periodically rotate the certificate and private key!